[enh] moving Controller class to ext.controllers.controller
