+++ sala_de_estudos_flask/ext/api/controllers/lessons_controller.py
@@ -1,6 +1,6 @@
 from http.client import NOT_FOUND
 
-from sala_de_estudos_flask.ext.api.controllers import Controller
+from sala_de_estudos_flask.ext.api.controllers.controller import Controller
 from sala_de_estudos_flask.ext.dao.lessons_dao import LessonsDAO
 
 
@@ -15,6 +15,6 @@ class LessonsController(Controller):
         return self.make_error(
             NOT_FOUND,
             "Not Found",
-            "It was not found lessons for this subject. "
+            f"It was not possible to find lessons for subject_id {subject_id}."
             "Try another subject."
         )
